Change target SOC entity's object
Update the name of the object which would be pulled from the API but nonexistent on e-up (MBB platform in API). With my minimal python skills I picked the only number returned by the API and used it instead of the nonexistent, to keep Home Assistant from crashing internally. The setup went through without any issues after this.